refactor : PR 리뷰 반영 1. 회원가입 관련 로직을 auth.controller, auth.service로 이동
2. switch 적용

3. update-user.dto를 PartialType 사용해 상속성 증가 및 추후 유지보수 간편화

4. User 엔티티에 Unique값을 추가하여 checkEmail 함수 제거
