+++ src/auth/auth.service.ts
@@ -1,9 +1,17 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { MailerService } from '@nestjs-modules/mailer';
+import { CreateUserDto } from 'src/dto/user/create-user.dto';
+import { UserService } from 'src/user/user.service';
+import { UserRepository } from 'src/repository/user.repository';
+import * as bcryptjs from 'bcryptjs';
 
 @Injectable()
 export class AuthService {
-  constructor(private readonly mailerService: MailerService) {}
+  constructor(
+    private readonly mailerService: MailerService,
+    private readonly userRepository: UserRepository,
+  ) {}
+
   async sendEmail(email: string) {
     const code: string = Math.round(Math.random() * 10000).toString();
     this.mailerService
@@ -21,5 +29,22 @@ export class AuthService {
         console.log(err);
       });
     return code;
+  }
+
+  async checkEmail(email: string) {
+    return await this.userRepository.checkEmail(email);
+  }
+
+  async signupWithEmail(createUserDto: CreateUserDto) {
+    Logger.log('Signup with Email Service Start');
+
+    Logger.log('Create Hashed Password Start');
+    const password: string = createUserDto.password;
+    const salt = bcryptjs.genSaltSync(10);
+    const hashedPassword = bcryptjs.hashSync(password, salt);
+    Logger.log('Create Hashed Password End');
+
+    await this.userRepository.signupWithEmail(createUserDto, hashedPassword);
+    Logger.log('Signup with Email Service End');
   }
 }
