+++ src/user/user.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from 'src/entity/user.entity';
 import { UserRepository } from 'src/repository/user.repository';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
